Atualizando a organização do codigo

+++ backend/src/main/java/com/pagamentos/projeto_programacao/voucher/DTO/DTOVoucher.java
@@ -2,7 +2,9 @@ package com.pagamentos.projeto_programacao.voucher.DTO;
 
 import com.pagamentos.projeto_programacao.voucher.Voucher;
 
-public record DTOVoucher(Long id, String cupom, int qt_used, float discount) {
+import java.math.BigDecimal;
+
+public record DTOVoucher(Long id, String cupom, int qt_used, BigDecimal discount) {
     public DTOVoucher(Voucher voucher) {
         this (
                 voucher.getId(),
